多重起動チェック処理を追加

+++ ha-batch/docker/aws_sqs_import.sh
@@ -20,6 +20,16 @@ echo "------------------------------------------------------------------------"
 # call common.sh
 . ./common.sh
 
+LOCK_FILE="/var/lock/aws_sqs_import.lock"
+mkdir -p "$(dirname "$LOCK_FILE")"
+
+exec 9>"${LOCK_FILE}"
+
+if ! flock -n 9; then
+  echo "[ERROR] awsSqsImportBatchJob is already running."
+  exit 50
+fi
+
 cd ${BASE_DIR} && docker compose -f docker-compose.yml -f docker-compose.${ENV}.yml run --rm ha-batch --spring.batch.job.name=awsSqsImportBatchJob
 
 echo "------------------------------------------------------------------------"
